requirements.txt
Django==2.2.6
djangorestframework==3.13.1
djoser==2.1.0
django-filter==21.1
